Creación de actividades con el archivo gpx.

+++ ApiServer/DTOs/Requests/ActivityDto.cs
@@ -1,5 +1,6 @@
 using System;
 using System.ComponentModel.DataAnnotations;
+using Microsoft.AspNetCore.Http;
 
 namespace ApiServer.DTOs.Requests
 {
@@ -8,13 +9,13 @@ namespace ApiServer.DTOs.Requests
         [Required]
         public string UserId { get; set; }
         [Required]
-        public float Distance { get; set; }
+        public decimal Distance { get; set; }
         [Required]
         public DateTime Duration { get; set; }
         [Required]
         public string Route { get; set; }
         [Required]
-        public float Altitude { get; set; }
+        public decimal Altitude { get; set; }
         [Required]
         public DateTime Start { get; set; }
         [Required]
